Add skeleton of game loop
Note the call to `free_all` with the temporary allocator. I'm taking directions from Karl Zylinski's excellent book on Odin to use the temporary allocator for everything allocated on a frame by frame basis.

If I remember correctly, context is always implicitly passed by default to new procedures I write, but I think it may be a best practice to add a context parameter. We'll see.