Add color buffer and color buffer texture then use them to draw colors to the screen
`startup` is officially ridiculous. I am going to break it up into multiple functions next commit, or I'm going to just pull the functionality from the function into the body of `main`.

I struggled with getting a `rawptr` from the color buffer but immediately after fixing it I got to see my yellow pixels drawn to the screen and it was very satisfying. Translating this to Odin is not too hard and it's cementing my understanding of Odin and lower level principles.