Move `startup` body into `main`
There was no real need to extract them to a function anyway. If I see the need in the future, I'll make multiple small functions. I can still use a neat shutdown function just as before.

Gustavo used globals for convenience in the course, so his use of a setup function made more sense and did lead to tidier code.