Implement rectangle drawing and migrate away from Rider
My mental model of using arrays in this way is not the best. I thought trying to draw out of bounds would cause problems but I tested it and it didn't. I don't know if I'm benefitting from Odin's bounds checking or if I was just wrong.

Going out of bounds with negative numbers is a problem, though. I'm okay with it. I consider it simply incorrect to supply negative numbers for x, y, width, or height of a rectangle.