Extract cube to global struct and rotate cube each frame
I caved on the global problem. It's okay for the cube because it's an example and I'll need to overhaul my code for general purpose use later anyway.

Now the cube rotates on all axes each frame. This feels like power in my hands. It looks like an old PS1 tech demo or something to me (albeit less impressive). It feels like magic. I am so excited to continue with graphics.