Import a .OBJ to render and write a custom .OBJ parser
My parser sucks but now that it exists I may iterate on it. This, to me, is the Zen of programming.

I had to synchronize with Gustavo a bit and stop tracking the tris directly but instead track faces (indices to vertices) on the mesh and then get the tris with those face vertices on the fly. I should be good now.

Next, with Gustavo's direction, I will see if there are any immediately obvious holes in my parser. It seems to work fine for now, though.