Implement back-face culling and load another cube for testing
I'm not sure the car looks right, but I'm going to assume it is some incompatibility in the data. Someone in the discussion forum mentioned that concave surfaces are not culled correctly by this algorithm, so maybe that's it.

Anyway, I tested with the cube and the back-faces are definitely being culled.