Convert `Projected_Triangle` and `Face` to structs, then store color info on them
This is a little less simple than my original implementation but I think it's just fine. There was only one place where I had to modify my logic a bit to account for not using arrays (the rest of the changes were procedure signatures and how to get at the data inside the objects).

I really want to use arrays wherever possible, but I think it makes logical sense to store color info on faces and projected triangles.